Fixed some typos.
Expanded and updated the section on floppies.

Rearranged section 2, moving the urls to the top where other sections can
refer to them. I have a section on installing from USB that needs this
reorg, but am not committing it yet since the necessary udebs are all stuck
in NEW.


Here's the USB section, so you'll see why the big changes:

2.3 USB memory stick

It's also possible to install from removable USB storage devices. For
example a USB keychain can make a handy Debian install medium you can take
with you anywhere. Your USB memory stick must have a DOS partition on it
for this method to work. Most come with such a partition by default.

To prepare your USB memory stick, download hd-media-initrd.gz and
save it to a file named "initrd.gz" on the USB device. Make sure that the
compressed image is not transparently decompressed by your browser.

You must also include a Debian netinst CD image on your memory stick. See
section 2.1 for details. Place the downloaded ISO on your memory stick;
any filename is ok as long as it ends in ".iso".

Some BIOSes can boot USB storage directly, and some cannot.

2.3.1 Booting directly from USB storage

The author of this section has not yet succeeded in booting directly from
USB storage, as all his hardware has too old a BIOS, but in theory you just
need to put a kernel on the USB device, and set up a boot loader. The
kernel can be gotten from the same place you'd get floppy images (see
section 2.2), and is named vmlinuz. If you succeed, please post details to
debian-boot@lists.debian.org so this section can be fixed.

2.3.2 Using USB storage and a boot floppy

The debian-installer can be booted off a single floppy, which will be able
to access your USB memory stick. To boot the installer this way, you will
need to put the bootfloppy-image.img on a floppy (see section 2.2).

Now boot from the floppy with the bootfloppy-image on it. It should detect
your USB device and mount the initrd.gz, which will in turn find the ISO.
After that things will proceed the same as if you had booted from CD.