Add a reversed Package/Provides mapping.
edos wasn't previously mentioning virtual packages. Since dose does that
now, build a mapping to perform provided/provider lookups. This lowers
the likelihood of having separate graphs when that shouldn't be the case.

An example can be seen with ia64 in wheezy (e.g. with the installed-based
and mounted-partitions virtual packages).