.mrconfig: Add ".git" suffix to Salsa URLs to avoid redirection
URLs of the form https://salsa.debian.org/<user>/<project> work for
both web browsers and git, but git is first redirected to the same URL
plus a ".git" suffix.  By default, git reports this redirection.  This
is a slight waste of resources and adds a fair amount of noise to the
"mr" output.

For an authenticated checkout using ssh, where the URLs are mapped to
the ssh protocol, it doesn't seem to make any difference whether the
".git" suffix is present.

So add the suffix where it's missing.