scripts/git-setup: Fix anon/auth detection if the script is re-run
"git remote -v" will follow the current URL mappings.  So if this
script is run on a repository with anonymous access to installer-team,
and the remote is then changed (in the repository config) to
authenticated, running the script again will still see it as having
anonymous access and won't update the URL mapping.

git remote doesn't seem to provide an option to ignore mappings, and
-c options can't be used to unset them either.  Use "git config
--local" to read the repository's remote URLs instead.