Revamp prepare-release-announce
Instead of crunching d-i logs, focus on the state of the archive when
d-i and installation images were built, using archive serial. This makes
it possible to look at udebs that aren't pulled during the build (but at
runtime instead).

This commit is a rebase+squash of rather old commits, isn't entirely
clean, but let's finally push that to master.